Convert TriviaQuizCreatorForm to TriviaQuizConfig
I want the player to be given the ability to retry a trivia quiz that
he/she just took. This is possible by storing a `TriviaQuizConfig` in
`UserDefaults`.
