+++ Tensai/Networking/TriviaQuizRequest.swift
@@ -3,8 +3,8 @@ import Foundation
 /// A request to retrieve trivia questions from the *Open Trivia Database*.
 struct TriviaQuizRequest: APIRequest {
     
-    func makeRequest(from data: TriviaQuizCreatorForm) -> URLRequest {
-        return URLRequest(url: data.apiURL!)
+    func makeRequest(from config: TriviaQuizConfig) -> URLRequest {
+        return URLRequest(url: config.apiURL!)
     }
     
     func parseResponse(data: Data) throws -> OpenTriviaDatabaseResponse {
