--- Tensai/View Models/TriviaQuizCreatorForm.swift
@@ -1,99 +1,0 @@
-import SwiftUI
-
-/// A form to retrieve trivia questions from the *Open Trivia Database*.
-///
-/// A URL to retrieve trivia questions is generated by selecting the category,
-/// difficulty level, question type, and number of questions.
-final class TriviaQuizCreatorForm: ObservableObject {
-    
-    /// The index of the currently selected category.
-    @Published var categoryIndex = 0
-    
-    /// The index of the currently selected difficulty level.
-    @Published var difficultyIndex = 0
-    
-    /// The index of the currently selected question type.
-    @Published var questionTypeIndex = 0
-    
-    /// The index of the currently selected number of questions.
-    @Published var questionCountIndex = 0
-    
-    /// The list of all possible categories.
-    static let allCategories = [
-        "Any Category",
-        "General Knowledge",
-        "Books",
-        "Film",
-        "Music",
-        "Musicals & Theater",
-        "Television",
-        "Video Games",
-        "Board Games",
-        "Science & Nature",
-        "Computers",
-        "Mathematics",
-        "Mythology",
-        "Sports",
-        "Geography",
-        "History",
-        "Politics",
-        "Art",
-        "Celebrities",
-        "Animals",
-        "Vehicles",
-        "Comics",
-        "Gadgets",
-        "Anime & Manga",
-        "Cartoons"
-    ]
-    
-    /// The list of all possible difficulty levels.
-    static let allDifficulties = [
-        "Any Difficulty",
-        "Easy",
-        "Medium",
-        "Hard"
-    ]
-    
-    /// The list of all possible question types.
-    static let allQuestionTypes = [
-        "Any Type",
-        "Multiple-Choice",
-        "True/False"
-    ]
-    
-    /// The list of all possible numbers of questions.
-    static let allQuestionCounts = (2 ... 10).map { $0 * 5 }
-    
-    /// The URL to retrieve trivia questions from the *Open Trivia Database*.
-    var apiURL: URL? {
-        
-        var components = URLComponents()
-        components.scheme = "https"
-        components.host = "opentdb.com"
-        components.path = "/api.php"
-        
-        let questionCount = TriviaQuizCreatorForm
-            .allQuestionCounts[questionCountIndex]
-        var queryItems = [
-            URLQueryItem(name: "amount", value: String(questionCount))
-        ]
-        if categoryIndex > 0 {
-            queryItems.append(
-                URLQueryItem(name: "category", value: String(categoryIndex + 8))
-            )
-        }
-        if difficultyIndex > 0 {
-            let value = TriviaQuizCreatorForm.allDifficulties[difficultyIndex]
-                .lowercased()
-            queryItems.append(URLQueryItem(name: "difficulty", value: value))
-        }
-        if questionTypeIndex > 0 {
-            let value = (questionTypeIndex == 1) ? "multiple" : "boolean"
-            queryItems.append(URLQueryItem(name: "type", value: value))
-        }
-        components.queryItems = queryItems
-        
-        return components.url
-    }
-}
