+++ TensaiTests/Networking Tests/TriviaQuizRequestTests.swift
@@ -6,12 +6,12 @@ final class TriviaQuizRequestTests: XCTestCase {
     private let request = TriviaQuizRequest()
     
     func testMakingURLRequest() {
-        let form = TriviaQuizCreatorForm()
-        form.categoryIndex = 1
-        form.difficultyIndex = 1
-        form.questionTypeIndex = 1
-        form.questionCountIndex = 8
-        let urlRequest = request.makeRequest(from: form)
+        var config = TriviaQuizConfig()
+        config.categoryIndex = 1
+        config.difficultyIndex = 1
+        config.questionTypeIndex = 1
+        config.questionCountIndex = 8
+        let urlRequest = request.makeRequest(from: config)
         let expectedURL = URL(
             string: [
                 "https://opentdb.com/api.php?amount=50",
